+++ mentalhealth/src/main/java/com/health/mentalhealth/domain/service/CalendarService.java
@@ -1,5 +1,6 @@
 package com.health.mentalhealth.domain.service;
 
+import com.health.mentalhealth.application.exception.RequestException;
 import com.health.mentalhealth.domain.persistence.entity.Calendar;
 import com.health.mentalhealth.domain.persistence.ports.in.ICalendarUseCase;
 import com.health.mentalhealth.infrastructure.repository.CalendarRepository;
@@ -24,11 +25,13 @@ public class CalendarService implements ICalendarUseCase {
 
     @Override
     public void deleteCalendar(Long id) {
+        if(!calendarUseCase.existsById(id)) throw new RequestException("404","calendar doesn't exist :(");
         calendarUseCase.deleteById(id);
     }
 
     @Override
     public Optional<Calendar> getCalendar(Long id) {
+        if(!calendarUseCase.existsById(id)) throw new RequestException("404","calendar doesn't exist :(");
         return calendarUseCase.findById(id);
     }
