+++ mentalhealth/src/main/java/com/health/mentalhealth/domain/service/RoutinesService.java
@@ -1,5 +1,6 @@
 package com.health.mentalhealth.domain.service;
 
+import com.health.mentalhealth.application.exception.RequestException;
 import com.health.mentalhealth.domain.persistence.entity.Routines;
 import com.health.mentalhealth.domain.persistence.ports.in.IRoutinesUseCase;
 import com.health.mentalhealth.infrastructure.repository.RoutinesRepository;
@@ -22,11 +23,13 @@ public class RoutinesService implements IRoutinesUseCase {
 
     @Override
     public void deleteRoutines(Long id) {
+        if(!routinesUseCase.existsById(id)) throw new RequestException("404","routine doesn't exist :(");
         routinesUseCase.deleteById(id);
     }
 
     @Override
     public Optional<Routines> getRoutinesById(Long id) {
+        if(!routinesUseCase.existsById(id)) throw new RequestException("404","routine doesn't exist :(");
         return routinesUseCase.findById(id);
     }
 
@@ -36,6 +39,7 @@ public class RoutinesService implements IRoutinesUseCase {
     }
 
     public List<Routines> findByName(String name){
+        if(routinesUseCase.findRoutinesByName(name)==null) throw new RequestException("404","routine with name: " + name+ " doesn't exist :(");
         return (List<Routines>) routinesUseCase.findRoutinesByName(name);
     }
 }
