+++ mentalhealth/src/main/java/com/health/mentalhealth/domain/service/UserService.java
@@ -1,4 +1,5 @@
 package com.health.mentalhealth.domain.service;
+import com.health.mentalhealth.application.exception.RequestException;
 import com.health.mentalhealth.domain.persistence.entity.UserEntity;
 import com.health.mentalhealth.infrastructure.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,21 +16,29 @@ public class UserService {
 
 
     public UserEntity createUser(UserEntity user) {
+        if(user.getEmail().isEmpty() ||  user.getPassword().isEmpty()) throw new RequestException("401", "email or password required");
         return userRepository.save(user);
     }
 
 
     public void deleteUser(Long id) {
+        if(!userRepository.existsById(id)) throw new RequestException("404","user doesn't exist :(");
         userRepository.deleteById(id);
     }
 
 
     public Optional<UserEntity> getUser(Long id) {
+        if(!userRepository.existsById(id)) throw new RequestException("404","user doesn't exist :(");
         return userRepository.findById(id);
     }
 
 
     public List<UserEntity> getAllUser() {
         return (List<UserEntity>) userRepository.findAll();
+    }
+
+    public UserEntity updateUser(UserEntity user){
+        if(!userRepository.existsById(user.getId())) throw new RequestException("404","user doesn't exist :(");
+        else return userRepository.save(user);
     }
 }
